+++ lib/devise/hooks/forgetable.rb
@@ -4,6 +4,6 @@
 # This avoids forgetting deleted users.
 Warden::Manager.before_logout do |record, warden, options|
   if record.respond_to?(:forget_me!)
-    Devise::Controllers::Rememberable::Proxy.new(warden).forget_me(record)
+    Devise::Hooks::Proxy.new(warden).forget_me(record)
   end
 end